qcacmn: Add service bit support for tx and rx time in channel stats
Add service bit support for channel tx and rx time in channel stats.
This service bit helps in not interpreting the default values for
tx_time and rx_time when new driver version is used with old firmware
where this tx_time and rx_time support is not present.

Change-Id: Icc7668f5842677ac6ef48176568301753e217ea7
CRs-Fixed: 2751500
Signed-off-by: chrisl7 <wandersonrodriguesf1@gmail.com>